Adds 'turtle_02' using the loop method to create the turtles:
This loop design pattern to create the turtle objects from fellow
FutureLearn learner, Ilia Shevelev, many thanks!
The race feedback can be viewed at the command line as print statements
showing final postions and winner.
The 'PLAYERS' list is purely for running a 'for loop' on each turtle
to reset the canvas without having to exit the runtime env.